requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
fpdf==1.7.2
frozenlist==1.5.0
google_search_results==2.4.2
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.31
lxml==5.3.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
ollama==0.4.8
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
zstandard==0.23.0
